+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `Value::push` method to add values to the end of lists.
+
 ### Changed
 
 - Allow index to be equal to length in Value::insert